% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAC_core_exported.R
\name{calculate_TOP_ART_score}
\alias{calculate_TOP_ART_score}
\title{calcuation of TOP-ART score}
\usage{
calculate_TOP_ART_score(
  SAMPLE_ID = NULL,
  seqMethod,
  HRD_LOH,
  LST,
  purity,
  ploidy,
  sex,
  bamDna,
  bamRna = NULL,
  vcf = NULL,
  somCna,
  somSmallVars = NULL,
  germSmallVars = NULL,
  geneModel,
  haploBlocks = NULL,
  yapsaTable = NULL,
  FILE_REFGEN = NULL,
  TOTAL_SNVS_YAPSA = NULL,
  GR_SOM_SNV_YAPSA = NULL,
  GR_FULL_VARS = NULL,
  PRINT_OUTPUT = F,
  OUTPUT_DIR = NULL,
  MAN_SOM_GENES = NULL,
  MAN_GERM_GENES = NULL,
  DEBUG = F,
  FILTER_SOM_SMALL_VARS = T,
  showReadDetail = F
)
}
\arguments{
\item{SAMPLE_ID}{optional ID of sample can be added}

\item{seqMethod}{charcter: either "WGS" or "WES"}

\item{HRD_LOH}{number of LOH events (wait for final text in manuscript)}

\item{LST}{number of large scale state transition events}

\item{purity}{purity of the sample (numeric value between 0 and 1 indicating 
the fraction of relevant sample with control/unrelevant tissue)}

\item{ploidy}{ploidy of the sample (numeric value)}

\item{sex}{sex of the sample (character: "male", "female", "m", "f")}

\item{bamDna}{path to bam-file}

\item{bamRna}{optional; path to rna file (bam format)}

\item{vcf}{character; path to variant call file (.vcf.gz format). 
Will be used (if provided)
for extended SNP phasing if variants on the same gene are too far away from
each other for direct haplotype phasing}

\item{somCna}{GRanges object containing all genomic regions with annotated 
total copynumber and cna_type as metadata columns. The total-copynumber 
column should be named "tcn" but also some other commonly used names. 
It should contain numeric values or characters that can be converted to 
numeric values. The cna_type column must contain the information about 
loss of heterozygosity (LOH). Therefore the term "LOH" must be explicitely 
mentioned in the column. If a genomic region is not present in the object, 
it will be taken as heterozygous with neutral TCN of 2.}

\item{somSmallVars}{GRanges object containing all somatic small 
variants (SNV and INDEL).
Required metadata columns are reference base (ref/REF), 
alternative base (alt/ALT),
annotation of the gene name (gene/GENE) and the allele-frequency (af/AF). 
If the object is not provided the tool assumes there are no somatic small 
variants.}

\item{germSmallVars}{GRanges object containing all germline small 
variants (SNV and INDEL).
Required metadata columns are reference base (ref/REF), alternative 
base (alt/ALT),
annotation of the gene name (gene/GENE) and the allele-frequency (af/AF)
If the object is not provided the tool assumes there are no germline small 
variants.}

\item{geneModel}{GRanges object containing the gene-annoattion of 
the used reference genome with metadata column of the gene name (gene)}

\item{haploBlocks}{GRanges object containing haploblocks. Haploblocks are
defined as genomic regions in which SNPs are phased to a specific allele.
For example a haploblock could be chr1:1000-10000. This would mean that every
genotype annotation in the format "1|0" or "0|1" of a SNP in this region will 
be used to phase somatic variants and define their genotype}

\item{yapsaTable}{table conatining the mutational signatures}

\item{FILE_REFGEN}{currently not usable.. required if YAPSA is incorporated}

\item{TOTAL_SNVS_YAPSA}{number of SNVs used to create yapsa signatures}

\item{GR_SOM_SNV_YAPSA}{currently not usable.. required if YAPSA is incorporated}

\item{GR_FULL_VARS}{currently not usable.. required if YAPSA is incorporated}

\item{PRINT_OUTPUT}{should output be printed into the concole?}

\item{OUTPUT_DIR}{output directory}

\item{MAN_SOM_GENES}{if somatic genes of relevance deviate from original TOP-ART gene set}

\item{MAN_GERM_GENES}{if germline genes of relevance deviate from original TOP-ART gene set}

\item{DEBUG}{debugging output}

\item{FILTER_SOM_SMALL_VARS}{should somatic variants be filtered for functional regions? - as defined for TOP-ART}

\item{showReadDetail}{exports an seperate table containg read info for reads used for phasing (see ZygosityPredictor)}
}
\description{
calcuation of TOP-ART score
}
\examples{
 
  
## we use the example dataset from the Bioconductor package ZygosiytPredictor
library(ZygosityPredictor)

bamfile <- system.file("extdata", "ZP_example.bam", 
                       package = "ZygosityPredictor")
                       
## load ZygosityPredictor variant datasets
data("GR_GERM_SMALL_VARS")
data("GR_SCNA")
data("GR_SOM_SMALL_VARS")
data("GR_GENE_MODEL")

## create mutational signatures table as it is produced by bioconductor package
## YAPSA
yapsaTable <- data.frame(
  sig=c("AC1","AC3","AC8"),
  exposure=c(4566, 2123, 598),
  lower=c(3899, 1902, 430),
  upper=c(5698, 2455, 789)
)

library(TopArtCalc)
## calculate TOP-ART score
topart_run <- calculate_TOP_ART_score(
  seqMethod="WGS",
  HRD=10,
  LST=4,
  purity=0.9,
  ploidy=2,
  sex="female",
  ## files
  bamDna=bamfile,
  ## granges
  somCna=GR_SCNA,   # meta cols required: tcn, cna_type
  somSmallVars=GR_SOM_SMALL_VARS,   # meta cols required: gene, af, ref, alt
  germSmallVars=GR_GERM_SMALL_VARS,  # meta cols required: gene, af, ref, alt, ACMG_class
  geneModel=GR_GENE_MODEL,     # meta cols required: gene
  yapsaTable=yapsaTable
  )
}
