% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAC_core_general.R
\name{estimate_germline_score}
\alias{estimate_germline_score}
\title{calc germline score}
\usage{
estimate_germline_score(germSmallVars, GERM_TA_GENES)
}
\value{
A List.
}
\description{
calc germline score
}
\keyword{internal}
