% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAC_core_general.R
\name{estimate_rearrangement_score}
\alias{estimate_rearrangement_score}
\title{calc rearrangement score}
\usage{
estimate_rearrangement_score(HRD_LOH, LST)
}
\value{
A List.
}
\description{
calc rearrangement score
}
\keyword{internal}
