% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAC_core_general.R
\name{estimate_somatic_score}
\alias{estimate_somatic_score}
\title{calc somatic score}
\usage{
estimate_somatic_score(
  somCna,
  somSmallVars,
  purity,
  sex,
  gr_germ,
  DEBUG,
  SOM_TA_GENES,
  GR_SOM_TA_GENES,
  bamDna,
  bamRna,
  PLOIDY,
  FILE_VCF,
  haploBlocks,
  showReadDetail,
  OUTPUT_DIR
)
}
\value{
A List.
}
\description{
calc somatic score
}
\keyword{internal}
