% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAC_core_exported.R
\name{load_germline_topart_genes}
\alias{load_germline_topart_genes}
\title{load genes of germline TOP-ART criterion as character vector}
\usage{
load_germline_topart_genes()
}
\value{
A character vector.
}
\description{
load genes of germline TOP-ART criterion as character vector
}
\examples{
library(TopArtCalc)

germline_topart_genes <- load_germline_topart_genes()
}
